+++ styles/search/result-list.module.css
@@ -1,0 +1,22 @@
+.select_container {
+  z-index: 1000;
+  max-width: 480px;
+  position: fixed;
+  display: flex;
+  width: 100%;
+  background-color: #fbfbfb;
+  top: 55px;
+  padding: 0 16px;
+  flex-direction: row-reverse;
+}
+
+.thumbnail_container {
+  padding: 0 16px;
+}
+
+.thumbnail_row {
+  margin: 0;
+  padding: 0;
+  overflow: auto;
+  margin-top: 112px;
+}
