+++ styles/search/search-header.module.css
@@ -19,10 +19,6 @@
   background: #fff;
 }
 
-.header:active {
-  transform: translate(100%);
-}
-
 .search_container {
   width: 100%;
   padding: 0 16px;
@@ -54,27 +50,4 @@
 
 .input_focused {
   background: #e9e9e9;
-}
-/*select box*/
-.select_container {
-  z-index: 1000;
-  max-width: 480px;
-  position: fixed;
-  display: flex;
-  width: 100%;
-  background-color: #fbfbfb;
-  top: 55px;
-  padding: 0 16px;
-  flex-direction: row-reverse;
-}
-
-.thumbnail_container {
-  padding: 0 16px;
-}
-
-.thumbnail_row {
-  margin: 0;
-  padding: 0;
-  overflow: auto;
-  margin-top: 112px;
 }
